refactor: Introduce tools.add_mpirun_script() and enable hyperthreading
apart from the refactoring, the main change that this commit introduces
is that `mpirun` will be invoked with `--use-hwthread-cpus` by default.
This change seemsto be necessary to let `mpirun` run on Ryzen CPUs.
Some relevant links are mentioned at #34.

This introduces an additional primary dependency, `psutil`. To be precise,
`psutil` was already a secondary dependency pulled by dask/distributed
so this does not really add anything extra. Nevertheless, with this commit
we promote it to a primary dependency.

Fixes #34