tests: Reduce the number of tests that download SRTM over the network
When running the tests (at least locally) the majority of the time is being
spent in downloading SRTM. With this commit we reduce the number of tests
that download the DEM data over the network (i.e. we do it with just a
single spatial window instead of multiple ones).

This should speed up the main tests by a fair bit (e.g. from ~5 minutes
down to less than 2 minutes). The tests that require `--runschism` should
not be affected.