meteo: Remove `meteo_engine`
We add a dedicated function for dispatching the `meteo_source`. The
function is relatively complicated but we also add extensive tests for it.

The only refernces to `meteo_engine` that remain is in the Tutorial. I am
not touching those files because @brey is working on them.

The tests for the `from_url()` function are marked with XFAIL because
there is some problem with SSL and the server we currently use. As soon
as it is fixed we should remove the marks:

    [SSL: UNSAFE_LEGACY_RENEGOTIATION_DISABLED]

Fixes #39